refactor: divided roi selector into components for ease of maintenance
