+++ package.json
@@ -11,7 +11,7 @@
     "build:viewer": "pnpm --filter vizarr build",
     "build:roi-selector": "pnpm --filter @biongff/roi-selector build",
     "build:app": "pnpm --filter app build",
-    "build": "pnpm build:viewer && (pnpm build:roi-selector || true) && pnpm build:app"
+    "build": "pnpm build:viewer && pnpm build:roi-selector && pnpm build:app"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
