+++ viewer/src/state.ts
@@ -2,7 +2,13 @@ import { type Atom, atom } from "jotai";
 import { atomFamily, splitAtom, waitForAll } from "jotai/utils";
 import { RedirectError, rethrowUnless } from "./utils";
 
-import type { Deck, Layer } from "deck.gl";
+import type { Layer } from "deck.gl";
+
+/** Plain-data snapshot of the deck.gl canvas dimensions. */
+export interface ViewportSize {
+  width: number;
+  height: number;
+}
 import type { PrimitiveAtom } from "jotai";
 import type { AtomFamily } from "jotai/vanilla/utils/atomFamily";
 import type { Matrix4 } from "math.gl";
@@ -208,7 +214,7 @@ export interface Redirect {
 }
 export const redirectObjAtom = atom<Redirect | null>(null);
 
-export const viewportAtom = atom<Deck | null>(null);
+export const viewportAtom = atom<ViewportSize | null>(null);
 
 export const sourceInfoAtom = atom<WithId<SourceData>[]>([]);
